Initial commit
This is a project that serves as a vegan friendly travel guide to Knoxville.  I listed out restaurants, food trucks, resources, and events.

Images used in the project were obtained through Wikimedia Commons and some through permission of the owners.

I used RecyclerView and CardViews to manage my content, and this was my first time using a NavigationDrawer.